refactor: :recycle: move app.data.cache module to app.core.cache and update import statements
