+++ app/api/api_v1/endpoints/blocks.py
@@ -11,7 +11,7 @@ from app.api.api_v1.dependencies import (
 )
 from app.api.api_v1.pagination import paginate_search_results
 from app.config.api_settings import get_settings
-from app.data.cache import cached_data
+from app.core.cache import cached_data
 
 router = APIRouter()
