+++ app/api/api_v1/endpoints/characters.py
@@ -12,7 +12,7 @@ from app.api.api_v1.dependencies import (
 from app.api.api_v1.dependencies.filter_param_matcher import filter_param_matcher
 from app.api.api_v1.endpoints.util import get_character_details
 from app.api.api_v1.pagination import paginate_search_results
-from app.data.cache import cached_data
+from app.core.cache import cached_data
 from app.data.encoding import get_codepoint_string
 from app.db.session import DBSession, get_session
 from app.docs.dependencies.custom_parameters import (
