+++ app/api/api_v1/endpoints/planes.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Path
 
 import app.db.models as db
 from app.config.api_settings import get_settings
-from app.data.cache import cached_data
+from app.core.cache import cached_data
 
 router = APIRouter()
