+++ app/tests/test_character_endpoints/test_get_unicode_character_details/test_get_character_details.py
@@ -3,7 +3,7 @@ from functools import reduce
 
 import pytest
 
-from app.data.cache import cached_data
+from app.core.cache import cached_data
 from app.data.encoding import get_uri_encoded_value
 from app.db.character_props import PROPERTY_GROUPS
 from app.db.procs.get_char_details import get_prop_groups
